+++ cocoatouch/src/main/java/org/robovm/cocoatouch/foundation/NSWritingDirection.java
@@ -22,8 +22,8 @@ public enum NSWritingDirection implements ValuedEnum {
     LeftToRight(0),
     RightToLeft(1);
 
-    private final int n;
+    private final long n;
 
-    private NSWritingDirection(int n) { this.n = n; }
-    public int value() { return n; }
+    private NSWritingDirection(long n) { this.n = n; }
+    public long value() { return n; }
 }
