+++ cocoatouch/src/main/java/org/robovm/cocoatouch/glkit/GLKViewDrawableColorFormat.java
@@ -23,11 +23,11 @@ public enum GLKViewDrawableColorFormat implements ValuedEnum {
     RGBA8888(0),
     RGB565(1);
 
-    private final int n;
+    private final long n;
 
-    private GLKViewDrawableColorFormat(int n) { this.n = n; }
-    public int value() { return n; }
-    public static GLKViewDrawableColorFormat fromValue(int n) {
+    private GLKViewDrawableColorFormat(long n) { this.n = n; }
+    public long value() { return n; }
+    public static GLKViewDrawableColorFormat fromValue(long n) {
         for (GLKViewDrawableColorFormat v : values()) {
             if (n == v.value()) {
                 return v;
