+++ cocoatouch/src/main/java/org/robovm/cocoatouch/uikit/UICollectionViewScrollDirection.java
@@ -21,8 +21,8 @@ public enum UICollectionViewScrollDirection implements ValuedEnum {
     Vertical(0),
     Horizontal(1);
 
-    private final int n;
+    private final long n;
 
-    private UICollectionViewScrollDirection(int n) { this.n = n; }
-    public int value() { return n; }
+    private UICollectionViewScrollDirection(long n) { this.n = n; }
+    public long value() { return n; }
 }
